feat(website): add custom font face declarations
