+++ website/app/components/preview/Install.tsx
@@ -1,16 +1,6 @@
-import {
-	Badge,
-	Grid,
-	Group,
-	Tabs,
-	Text,
-	Title,
-	UnstyledButton,
-} from '@mantine/core';
+import { Grid, Group, Tabs, Text, Title, UnstyledButton } from '@mantine/core';
 import { Link } from '@remix-run/react';
-import { useState } from 'react';
 
-import { Code } from '@/components/code/Code';
 import { PackageManagerCode } from '@/components/code/PackageManagerCode';
 import { IconExternal } from '@/components/icons';
 import globalClasses from '@/styles/global.module.css';
@@ -19,14 +9,7 @@ import type { Metadata, VariableData } from '@/utils/types';
 import { CarbonAd } from '../CarbonAd';
 import { InfoWrapper } from './Info';
 import classes from './Install.module.css';
-
-const categoryMap: Record<string, string> = {
-	'sans-serif': 'sans-serif',
-	serif: 'serif',
-	monospace: 'monospace',
-	handwriting: 'cursive',
-	display: 'system-ui',
-};
+import { InstallCode } from './InstallCode';
 
 interface InstallProps {
 	metadata: Metadata;
@@ -35,69 +18,6 @@ interface InstallProps {
 }
 
 const Variable = ({ metadata, variable }: InstallProps) => {
-	const [isActive, setActive] = useState<Record<string, boolean>>({
-		wght: true,
-	});
-
-	// Remove ital from active axes and mark separate ital flag as true
-	const activeAxes = Object.keys(isActive).filter((axis) => axis !== 'ital');
-	const isItal = isActive.ital;
-
-	// Determine if it is a standard axis e.g. only contains wght, wdth, slnt, opsz or ital
-	const isStandard = activeAxes.every((axis) =>
-		['wght', 'wdth', 'slnt', 'opsz'].includes(axis),
-	);
-
-	const importComment =
-		metadata.weights.length === 1
-			? `// Supports only weight ${metadata.weights[0]}\n`
-			: `// Supports weights ${metadata.weights[0]}-${
-					metadata.weights.at(-1) ?? 400
-			  }\n`;
-
-	const generateImports = () => {
-		if (activeAxes.length === 1 && isActive.wght) {
-			if (isItal)
-				return `import '@fontsource-variable/${metadata.id}/wght-italic.css';`;
-			return `import '@fontsource-variable/${metadata.id}';`;
-		}
-
-		// If it is only wght and another axes, return axes.css
-		if (activeAxes.length === 2 && isActive.wght) {
-			const selected =
-				activeAxes.find((axis) => axis !== 'wght')?.toLowerCase() ?? 'wght';
-			if (isItal)
-				return `import '@fontsource-variable/${metadata.id}/${selected}-italic.css';`;
-			return `import '@fontsource-variable/${metadata.id}/${selected}.css';`;
-		}
-
-		// If the selected axes is within standard, only return standard
-		if (isStandard)
-			return `import '@fontsource-variable/${metadata.id}/standard.css';`;
-
-		// If the selected axes is not within standard, return full
-		return `import '@fontsource-variable/${metadata.id}/full.css';`;
-	};
-
-	const handleActive = (value: string | number) => {
-		setActive((prev) => {
-			if (prev[value]) {
-				// eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-				delete prev[value];
-				return {
-					...prev,
-					wght: true,
-				};
-			}
-
-			return {
-				...prev,
-				[value]: !prev[value],
-				wght: true,
-			};
-		});
-	};
-
 	return (
 		<>
 			<Title order={3} mt="xl" mb="md">
@@ -108,83 +28,12 @@ const Variable = ({ metadata, variable }: InstallProps) => {
 				command:
 			</Text>
 			<PackageManagerCode cmd={`@fontsource-variable/${metadata.id}`} />
-			<Title order={3} mt="xl" mb="md">
-				Import
-			</Title>
-			{variable && (
-				<Group>
-					{Object.keys(variable.axes).map((axis) => (
-						<Badge
-							key={axis}
-							className={classes.badge}
-							onClick={() => {
-								handleActive(axis);
-							}}
-							data-active={Boolean(isActive[axis])}
-						>
-							{axis}
-						</Badge>
-					))}
-				</Group>
-			)}
-			<Code language="jsx">{importComment + generateImports()}</Code>
-			<Title order={3} mt="xl" mb="md">
-				CSS
-			</Title>
-			<Text>
-				Include the CSS in your project by adding the following line to your
-				project:
-			</Text>
-			<Code language="css">
-				{`body {
-  font-family: '${metadata.family} Variable', ${
-		categoryMap[metadata.category] ?? 'sans-serif'
-	};
-}`}
-			</Code>
+			<InstallCode metadata={metadata} variable={variable} />
 		</>
 	);
 };
 
 const Static = ({ metadata }: InstallProps) => {
-	const [isActive, setActive] = useState<Record<string, boolean>>({
-		400: true,
-	});
-	const keys = Object.keys(isActive);
-	const handleActive = (value: string | number) => {
-		setActive((prev) => {
-			if (keys.length === 1 && prev[value]) return prev;
-			if (prev[value]) {
-				// eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-				delete prev[value];
-				return {
-					...prev,
-				};
-			}
-
-			return {
-				...prev,
-				[value]: !prev[value],
-			};
-		});
-	};
-
-	const [isItal, setIsItal] = useState(false);
-	const generateImports = () => {
-		let imports = '';
-		if (keys.length === 1 && isActive[400]) {
-			if (isItal) return `import '@fontsource/${metadata.id}/400-italic.css';`;
-			return `import '@fontsource/${metadata.id}';`;
-		}
-
-		for (const weight of keys) {
-			imports += isItal
-				? `import '@fontsource/${metadata.id}/${weight}-italic.css';\n`
-				: `import '@fontsource/${metadata.id}/${weight}.css';\n`;
-		}
-		return imports.trim();
-	};
-
 	return (
 		<>
 			<Title order={3} mt="xl" mb="md">
@@ -195,49 +44,7 @@ const Static = ({ metadata }: InstallProps) => {
 				command:
 			</Text>
 			<PackageManagerCode cmd={`@fontsource/${metadata.id}`} />
-			<Title order={3} mt="xl" mb="md">
-				Import
-			</Title>
-			<Group>
-				{metadata.weights.map((weight) => (
-					<Badge
-						key={weight}
-						className={classes.badge}
-						onClick={() => {
-							handleActive(weight);
-						}}
-						data-active={Boolean(isActive[weight])}
-					>
-						{weight}
-					</Badge>
-				))}
-				{metadata.styles.includes('italic') && (
-					<Badge
-						className={classes.badge}
-						onClick={() => {
-							setIsItal((prev) => !prev);
-						}}
-						data-active={isItal}
-					>
-						italic
-					</Badge>
-				)}
-			</Group>
-			<Code language="jsx">{generateImports()}</Code>
-			<Title order={3} mt="xl" mb="md">
-				CSS
-			</Title>
-			<Text>
-				Include the CSS in your project by adding the following line to your
-				project:
-			</Text>
-			<Code language="css">
-				{`body {
-  font-family: '${metadata.family}', ${
-		categoryMap[metadata.category] ?? 'sans-serif'
-	};
-}`}
-			</Code>
+			<InstallCode metadata={metadata} />
 		</>
 	);
 };
